Use thiserror to create TarError
TarError implements the From trait for common Deku, Io, Utf8Error, and
ParseIntError error types as well as a couple of erros that previously overloaded
std::io:Error. This also makes it easier for a user to handle the those
error cases we give instead of matching a string.

This also makes the calls to those functions all handled and given to
the user, instead of panic'ing uncontrollably.